some styling and "WINE" cat
added category for wines to be displayed in the "wine cellar"(if not
"import") or the shop (depending on availability).
couldn't import old fonts but found good replacements possibly. room for
improvement.
need to write logic to allow user to sort wines by
style/price/import/etc. that will be next.